scenarios/2_iot-hub-messaging/functions/requirements.txt
azure-functions==1.20.0
fastapi==0.111.0
azure-storage-blob==12.20.0
python-dotenv==1.0.1
openai==1.35.10
azure-iot-device==2.13.1
azure-iot-hub==2.6.1
